% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{getDataset}
\alias{getDataset}
\title{Acquisition of Data}
\usage{
getDataset()
}
\value{
data.frame
}
\description{
Get the dataset from the Opendata website of the Lombardy region.
}
\examples{
data <- getDataset()
}
