% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/year_injured.R
\name{year_injured}
\alias{year_injured}
\title{Mean of Injured People (Year)}
\usage{
year_injured(year, data = NULL)
}
\arguments{
\item{year}{numeric. Year (2010 to 2016) for which calculating the mean.}

\item{data}{data.frame.}
}
\value{
The mean of injured people in the choosen year divided by district.
}
\description{
Mean of injured people divided by district in a specific year.
}
\examples{
year_injured(2010)

}
