% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/district_dead.R
\name{district_dead}
\alias{district_dead}
\title{Mean of Dead People (District)}
\usage{
district_dead(data, district)
}
\arguments{
\item{data}{data.frame.}

\item{district}{character. District of Lombardy Region contained in the dataframe.}
}
\value{
The mean of dead people in the choosen district divided by year (2010 to 2016).
}
\description{
Mean of dead people divided by year in a specific district.
}
\examples{
district_dead(data, "MI")

}
