% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/district_dead.R
\name{district_dead}
\alias{district_dead}
\title{Mean of dead people divided by district}
\usage{
district_dead(district)
}
\arguments{
\item{district}{character. District of Lombardy Region contained in the dataframe.}
}
\value{
A plot with the mean of dead people in the choosen district divided by year (from 2010 to 2016).
}
\description{
The user choose the district and the function return the mean of dead people
divided by year.
}
\examples{
district_dead()

}
