% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/district_injured.R
\name{district_injured}
\alias{district_injured}
\title{Mean of Injured People (District)}
\usage{
district_injured(data, district)
}
\arguments{
\item{data}{data.frame.}

\item{district}{character. District of Lombardy Region contained in the dataframe.}
}
\value{
The mean of injured people in the choosen district divided by year (2010 to 2016).
}
\description{
Mean of injured people divided by year in a specific district.
}
\examples{
district_injured(data, "MI")

}
