% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/year_dead.R
\name{year_dead}
\alias{year_dead}
\title{Mean of Dead People (Year)}
\usage{
year_dead(data, year)
}
\arguments{
\item{data}{data.frame.}

\item{year}{numeric. Year (2010 to 2016) for which calculating the mean.}
}
\value{
The mean of dead people in the choosen year divided by district.
}
\description{
Mean of dead people divided by district in a specific year
}
\examples{
year_dead(data, 2010)

}
