% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/year_dead.R
\name{year_dead}
\alias{year_dead}
\title{Mean of dead people per district in a specific year}
\usage{
year_dead(year)
}
\arguments{
\item{year}{numeric. Year for which the user want to see the analysis (from 2010 to 2016).}
}
\value{
The mean of dead people in the choosen year divided by district.
}
\description{
The user choose the year and the function return the mean of dead people
divided by district
}
\examples{
year_dead(2010)

}
