% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/year_injured.R
\name{year_injured}
\alias{year_injured}
\title{Mean of injured people per district in a specific year}
\usage{
year_injured(year)
}
\arguments{
\item{year}{numeric. Year for which the user want to see the analysis (from 2010 to 2016).}
}
\value{
The mean of injured people in the choosen year divided by district.
}
\description{
The user choose the year and the function return the mean of injured people
divided by district
}
\examples{
year_injured(2010)

}
